refactor(app): extract LibraryPane + ActiveBook (no behaviour change) (app-21)
