+++ apps/android/lib/src/ui/library_home_screen.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 
 import '../data/companion_runtime.dart';
@@ -9,6 +7,7 @@ import '../domain/library_tree.dart';
 import '../domain/listen_progress.dart';
 import '../domain/paired_server.dart';
 import 'app_settings_screen.dart';
+import 'library_pane.dart';
 import 'player_screen.dart';
 
 /// Post-pairing home: a cover-art library grouped into author → series
@@ -43,6 +42,9 @@ class _LibraryHomeScreenState extends State<LibraryHomeScreen> {
   bool _offline = false;
   bool _connecting = false; // background server probe in flight
   String? _error;
+
+  /// Single source of truth for which book the detail pane shows (app-21).
+  final ActiveBook _activeBook = ActiveBook();
 
   @override
   void initState() {
@@ -196,7 +198,11 @@ class _LibraryHomeScreenState extends State<LibraryHomeScreen> {
     }
   }
 
-  Future<void> _openBook(String bookId, String title) async {
+  /// Task 6 temporary routing: always push PlayerScreen (current behaviour).
+  /// Task 7 replaces this with layout-aware routing (two-pane fills the pane
+  /// in place; single-pane keeps this push).
+  Future<void> _onSelect(String bookId, String title) async {
+    _activeBook.select(bookId, title: title);
     await Navigator.of(context).push(MaterialPageRoute(
       builder: (_) =>
           PlayerScreen(runtime: widget.runtime, bookId: bookId, title: title),
@@ -206,95 +212,26 @@ class _LibraryHomeScreenState extends State<LibraryHomeScreen> {
     if (mounted) await _refresh();
   }
 
-  void _open(LibraryBook book) => _openBook(book.bookId, book.title);
-
-  Widget _continueRail() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 12, 16, 4),
-          child: Text('Continue listening',
-              style: Theme.of(context)
-                  .textTheme
-                  .titleMedium
-                  ?.copyWith(fontWeight: FontWeight.bold)),
-        ),
-        SizedBox(
-          height: 156,
-          child: ListView(
-            scrollDirection: Axis.horizontal,
-            padding: const EdgeInsets.symmetric(horizontal: 8),
-            children: [for (final b in _continue) _shelfCard(b)],
-          ),
-        ),
-        const Divider(),
-      ],
-    );
-  }
-
-  Future<void> _confirmRemoveFromShelf(ShelfBook b) async {
-    final remove = await showModalBottomSheet<bool>(
-      context: context,
-      builder: (ctx) => SafeArea(
-        child: ListTile(
-          key: const Key('remove-from-shelf'),
-          leading: const Icon(Icons.remove_circle_outline),
-          title: const Text('Remove from Continue listening'),
-          onTap: () => Navigator.of(ctx).pop(true),
-        ),
-      ),
-    );
-    if (remove != true) return;
+  Future<void> _handleRemoveFromShelf(ShelfBook b) async {
     await widget.runtime.library.setBookHidden(b.bookId, true);
     widget.runtime.api.setShelfStatus(b.bookId, hidden: true).catchError((_) {});
     if (mounted) await _refresh();
   }
 
-  Widget _shelfCard(ShelfBook b) {
-    final path = _covers[b.bookId];
-    return InkWell(
-      key: Key('continue-${b.bookId}'),
-      onTap: () => _openBook(b.bookId, b.title),
-      onLongPress: () => _confirmRemoveFromShelf(b),
-      child: SizedBox(
-        width: 100,
-        child: Padding(
-          padding: const EdgeInsets.all(6),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              ClipRRect(
-                borderRadius: BorderRadius.circular(6),
-                child: SizedBox(
-                  width: 88,
-                  height: 112,
-                  child: path != null
-                      ? Image.file(File(path),
-                          fit: BoxFit.cover,
-                          errorBuilder: (_, _, _) => const Icon(Icons.menu_book))
-                      : ColoredBox(
-                          color: Theme.of(context).colorScheme.surfaceContainerHighest,
-                          child: const Icon(Icons.menu_book)),
-                ),
-              ),
-              const SizedBox(height: 4),
-              Flexible(
-                child: Text(b.title,
-                    maxLines: 2,
-                    overflow: TextOverflow.ellipsis,
-                    style: Theme.of(context).textTheme.bodySmall),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
+  Future<void> _handleRemoveDownload(LibraryBook book) async {
+    await widget.runtime.library.evictBook(book.bookId);
+    await _refresh();
+    if (mounted) {
+      ScaffoldMessenger.of(context)
+          .showSnackBar(SnackBar(content: Text('Removed "${book.title}".')));
+    }
   }
+
+  void _toggle(String key) => setState(
+      () => _collapsed.contains(key) ? _collapsed.remove(key) : _collapsed.add(key));
 
   @override
   Widget build(BuildContext context) {
-    final tree = buildLibraryTree(filterBooks(_books, _query));
     return Scaffold(
       appBar: AppBar(
         title: const Text('Library'),
@@ -335,261 +272,25 @@ class _LibraryHomeScreenState extends State<LibraryHomeScreen> {
           ),
         ],
       ),
-      body: Column(
-        children: [
-          Padding(
-            padding: const EdgeInsets.fromLTRB(12, 8, 12, 4),
-            child: TextField(
-              key: const Key('library-search'),
-              decoration: const InputDecoration(
-                prefixIcon: Icon(Icons.search),
-                hintText: 'Filter by author, series or title',
-                isDense: true,
-                border: OutlineInputBorder(),
-              ),
-              onChanged: (v) => setState(() => _query = v),
-            ),
-          ),
-          if (_loading)
-            const Expanded(child: Center(child: CircularProgressIndicator()))
-          else if (_error != null)
-            Expanded(
-              child: Center(
-                child: Padding(
-                  padding: const EdgeInsets.all(24),
-                  child: Text('Sync failed: $_error', key: const Key('library-error')),
-                ),
-              ),
-            )
-          else
-            Expanded(
-              child: ListView(
-                children: [
-                  if (_continue.isNotEmpty && _query.isEmpty) _continueRail(),
-                  for (final author in tree) ..._authorSection(author),
-                ],
-              ),
-            ),
-        ],
+      body: LibraryPane(
+        books: _books,
+        continueBooks: _continue,
+        covers: _covers,
+        progress: _progress,
+        totalSec: _totalSec,
+        listened: _listened,
+        collapsedKeys: _collapsed,
+        query: _query,
+        loading: _loading,
+        error: _error,
+        currentBookId: widget.runtime.player.currentBookId,
+        onQueryChanged: (v) => setState(() => _query = v),
+        onToggleCollapse: _toggle,
+        onSelect: (bookId, title) => _onSelect(bookId, title),
+        onDownload: _download,
+        onRemoveDownload: _handleRemoveDownload,
+        onRemoveFromShelf: _handleRemoveFromShelf,
       ),
     );
-  }
-
-  void _toggle(String key) => setState(
-      () => _collapsed.contains(key) ? _collapsed.remove(key) : _collapsed.add(key));
-
-  Widget _sectionHeader({
-    required String label,
-    required bool collapsed,
-    required VoidCallback onTap,
-    required double indent,
-    TextStyle? style,
-    String? trailing,
-  }) {
-    return InkWell(
-      onTap: onTap,
-      child: Padding(
-        padding: EdgeInsets.fromLTRB(indent, 10, 12, 6),
-        child: Row(
-          children: [
-            Icon(collapsed ? Icons.add_box_outlined : Icons.indeterminate_check_box_outlined,
-                size: 22),
-            const SizedBox(width: 8),
-            Expanded(child: Text(label, style: style)),
-            if (trailing != null)
-              Text(trailing, style: Theme.of(context).textTheme.bodySmall),
-          ],
-        ),
-      ),
-    );
-  }
-
-  List<Widget> _authorSection(AuthorGroup author) {
-    final aKey = 'author:${author.author}';
-    final aCollapsed = _collapsed.contains(aKey);
-    final bold = Theme.of(context)
-        .textTheme
-        .titleMedium
-        ?.copyWith(fontWeight: FontWeight.bold);
-    final bookCount =
-        author.series.fold<int>(0, (n, s) => n + s.books.length);
-    return [
-      _sectionHeader(
-        label: author.author,
-        collapsed: aCollapsed,
-        onTap: () => _toggle(aKey),
-        indent: 12,
-        style: bold,
-        trailing: '$bookCount',
-      ),
-      const Divider(height: 1),
-      if (!aCollapsed)
-        for (final series in author.series)
-          ..._seriesSection(author.author, series),
-    ];
-  }
-
-  List<Widget> _seriesSection(String author, SeriesGroup series) {
-    if (series.series.isEmpty) {
-      return [for (final book in series.books) _bookTile(book)];
-    }
-    final sKey = 'series:$author/${series.series}';
-    final sCollapsed = _collapsed.contains(sKey);
-    return [
-      _sectionHeader(
-        label: series.series,
-        collapsed: sCollapsed,
-        onTap: () => _toggle(sKey),
-        indent: 28,
-        style: Theme.of(context).textTheme.labelLarge,
-        trailing: '${series.books.length}',
-      ),
-      if (!sCollapsed) for (final book in series.books) _bookTile(book),
-    ];
-  }
-
-  Widget _bookTile(LibraryBook book) {
-    final downloading = _progress.containsKey(book.bookId);
-    return ListTile(
-      key: Key('book-${book.bookId}'),
-      leading: _cover(book.bookId),
-      title: Text(book.title),
-      subtitle: _subtitleWidget(book),
-      isThreeLine: true,
-      onTap: (book.downloadState == BookDownloadState.downloaded ||
-              book.downloadState == BookDownloadState.updateAvailable)
-          ? () => _open(book)
-          : null,
-      trailing: downloading
-          ? _progressWidget(book.bookId)
-          : _action(book),
-    );
-  }
-
-  Widget _cover(String bookId) {
-    final path = _covers[bookId];
-    final child = path != null
-        ? Image.file(File(path), fit: BoxFit.cover,
-            errorBuilder: (_, _, _) => const Icon(Icons.menu_book))
-        : const Icon(Icons.menu_book);
-    return SizedBox(
-      width: 44,
-      height: 60,
-      child: ClipRRect(borderRadius: BorderRadius.circular(4), child: child),
-    );
-  }
-
-  Widget _action(LibraryBook book) {
-    switch (book.downloadState) {
-      case BookDownloadState.notDownloaded:
-        return IconButton(
-          key: Key('download-${book.bookId}'),
-          icon: const Icon(Icons.download),
-          tooltip: 'Download',
-          onPressed: () => _download(book),
-        );
-      case BookDownloadState.updateAvailable:
-      case BookDownloadState.downloaded:
-        // Downloaded (tap row to play) — menu offers update/remove.
-        return PopupMenuButton<String>(
-          key: Key('book-menu-${book.bookId}'),
-          onSelected: (v) {
-            if (v == 'play') _open(book);
-            if (v == 'update') _download(book);
-            if (v == 'remove') _removeDownload(book);
-          },
-          itemBuilder: (_) => [
-            const PopupMenuItem(value: 'play', child: Text('Play')),
-            if (book.downloadState == BookDownloadState.updateAvailable)
-              const PopupMenuItem(
-                  value: 'update', child: Text('Update (re-sync)')),
-            const PopupMenuItem(
-                value: 'remove', child: Text('Remove download')),
-          ],
-        );
-      case BookDownloadState.downloading:
-        return _progressWidget(book.bookId);
-    }
-  }
-
-  Future<void> _removeDownload(LibraryBook book) async {
-    if (widget.runtime.player.currentBookId == book.bookId) {
-      ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-          content: Text('Stop playback before removing this book.')));
-      return;
-    }
-    final ok = await showDialog<bool>(
-      context: context,
-      builder: (ctx) => AlertDialog(
-        title: const Text('Remove download?'),
-        content: Text(
-            'Deletes the downloaded audio for "${book.title}". You can re-download it later.'),
-        actions: [
-          TextButton(
-              onPressed: () => Navigator.pop(ctx, false),
-              child: const Text('Cancel')),
-          FilledButton(
-              onPressed: () => Navigator.pop(ctx, true),
-              child: const Text('Remove')),
-        ],
-      ),
-    );
-    if (ok != true) return;
-    await widget.runtime.library.evictBook(book.bookId);
-    await _refresh();
-    if (mounted) {
-      ScaffoldMessenger.of(context)
-          .showSnackBar(SnackBar(content: Text('Removed "${book.title}".')));
-    }
-  }
-
-  Widget _progressWidget(String bookId) {
-    final p = _progress[bookId] ?? '…';
-    return Row(mainAxisSize: MainAxisSize.min, children: [
-      Text(p),
-      const SizedBox(width: 8),
-      const SizedBox(
-          width: 18, height: 18, child: CircularProgressIndicator(strokeWidth: 2)),
-    ]);
-  }
-
-  Widget _subtitleWidget(LibraryBook book) {
-    final pos = formatSeriesPosition(book.seriesPosition);
-    final seriesLine = book.series.isEmpty
-        ? null
-        : '${book.series}${pos.isNotEmpty ? ' #$pos' : ''}';
-    final total = _totalSec[book.bookId];
-    final status = _statusLabel(book.downloadState) +
-        (total != null ? ' · ${formatDuration(total)}' : '');
-    final listened = _listened[book.bookId];
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        if (seriesLine != null) Text(seriesLine),
-        Text(status),
-        if (listened != null && listened > 0)
-          Padding(
-            padding: const EdgeInsets.only(top: 4, right: 12),
-            child: LinearProgressIndicator(
-              value: listened,
-              minHeight: 4,
-              backgroundColor: Theme.of(context).colorScheme.surfaceContainerHighest,
-            ),
-          ),
-      ],
-    );
-  }
-
-  String _statusLabel(BookDownloadState s) {
-    switch (s) {
-      case BookDownloadState.notDownloaded:
-        return 'Not downloaded';
-      case BookDownloadState.downloading:
-        return 'Downloading…';
-      case BookDownloadState.downloaded:
-        return 'Downloaded · tap to listen';
-      case BookDownloadState.updateAvailable:
-        return 'Update available since last sync';
-    }
   }
 }
